refactor: finalize gameplay composition cleanup
