+++ frontend/src/features/gameplay/session/model/session.types.ts
@@ -1,44 +1,15 @@
 import type { Cell } from "@/features/gameplay/canvas";
-
-export interface RoundInfoState {
-  roundId: number | null;
-  roundNumber: number | null;
-  roundDurationSec: number | null;
-  totalRounds: number;
-  formattedGameEndTime: string | null;
-  remainingSeconds: number | null;
-  formattedRemainingTime: string | null;
-  isRoundExpired: boolean;
-}
-
-export interface VoteSessionState {
-  votes: Record<string, number>;
-  remaining: number | null;
-  votingCellIds: Set<number>;
-  topColorMap: Map<number, string>;
-}
-
-export interface CanvasSessionState {
-  canvasId: number | null;
-  gridX: number;
-  gridY: number;
-  cells: Cell[];
-  selectedCell: Cell | null;
-  previewColor: string | null;
-}
-
-export interface SessionStatusState {
-  loading: boolean;
-  error: string | null;
-  gameEnded: boolean;
-  canvasReady: boolean;
-}
+import type {
+  RoundInfo as RoundData,
+  RoundTimer,
+} from "@/features/gameplay/round/model/round.types";
 
 export interface SessionBootstrapResult {
   canvasId: number;
   gridX: number;
   gridY: number;
   cells: Cell[];
-  round: RoundInfoState;
+  round: RoundData | null;
+  timer: RoundTimer | null;
   remaining: number | null;
 }
